+++ app/ids-notification-banner/index.js
@@ -1,0 +1,3 @@
+import IdsNotificationBanner from '../../src/ids-notification-banner/ids-notification-banner';
+import IdsCard from '../../src/ids-card/ids-card';
+import IdsListView from '../../src/ids-list-view/ids-list-view';